Conflicts:
	src/game/Player.cpp
	src/game/Player.h
	src/game/Unit.cpp

+++ src/game/LootMgr.cpp
@@ -166,7 +166,7 @@ void LootStore::LoadLootTable()
         Verify();                                           // Checks validity of the loot store
 
         sLog.outString();
-        sLog.outString( ">> Loaded %u loot definitions (%d templates)", count, m_LootTemplates.size());
+        sLog.outString( ">> Loaded %u loot definitions (%lu templates)", count, (unsigned long)m_LootTemplates.size());
     }
     else
     {
